Mon, Sep 12, 2022, 12:31 AM +00:00

+++ hi.c
@@ -6,6 +6,7 @@ int main(void)
     char c2 = 'I';
     char c3 = '!';
 
-    printf("%i %i %i\n", c1, c2, c3);
+    printf("%i %i %i\n", (int) c1, (int) c2, (int) c3);
     // printf is printing the ASCII values of each of the letters
+    // we are converting the chars to ints
 }
